Refactor packages structure
