+++ core/src/main/scala/geotrellis/server/utils/Implicits.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 Azavea
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package geotrellis.server.utils
+
+import java.io.{PrintWriter, StringWriter}
+
+trait Implicits {
+  implicit class throwableExtensions[T <: Throwable](th: T) {
+    def stackTraceString: String = {
+      val writer = new StringWriter()
+      th.printStackTrace(new PrintWriter(writer))
+      writer.toString
+    }
+  }
+}
+
+object Implicits extends Implicits
